refactor for okhttp
