+++ app/src/main/java/com/example/a2chatAndroid/Network/Api/Retrofit/Service/firestoreApiService.kt
@@ -1,8 +1,8 @@
 package com.example.a2chatAndroid.Network.RetrofitApi.Service
 
 import android.util.Log
-import com.example.a2chatAndroid.Network.RetrofitApi.OnLobbyJoinRequest
-import com.example.a2chatAndroid.Network.RetrofitApi.RetroFitClient.apiService
+import com.example.a2chatAndroid.Network.Api.Retrofit.OnLobbyJoinRequest
+import com.example.a2chatAndroid.Network.Api.Retrofit.RetroFitClient.apiService
 
 
 //makes api request for /firestore/createLobby
